organize controllers responsible for authentication
