+++ holidaymaker/src/main/java/com/newton/holidaymaker/controllers/RegisterController.java
@@ -1,80 +1,43 @@
 package com.newton.holidaymaker.controllers;
 
+import java.security.Principal;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import org.apache.commons.validator.routines.EmailValidator;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCrypt;
-import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.ModelAndView;
+
 import com.newton.holidaymaker.models.User;
 import com.newton.holidaymaker.repositories.UserRepository;
 
 @RestController
-@RequestMapping("/auth")
-public class AuthenticationController extends PageControllerEssentials {
+public class RegisterController extends PageControllerEssentials implements PageControllerInterface {
+
+    @GetMapping("/register")
+    @Override
+    public ModelAndView run(HttpServletRequest req, HttpServletResponse res, Principal principal) {
+
+        // Redirect already-logged-in users to index
+        if(principal != null) {
+            redirect("/", res);
+            return null;
+        }
+
+        ModelAndView mv = initModelAndView("HolidayMaker | Register Account", "register", "register");
+        return mv;
+    }
 
     @Autowired
     private UserRepository userRepository;
-
-    /**
-    * Handles login requests.<br>
-    * a user session will be started upon successful login.
-    *
-    * @return status message; "err#<i>statusMessageSample</i>" or "success"
-    * */
-    @PostMapping("/login")
-    @ResponseBody
-    public ResponseEntity loginUser(@RequestBody User loginForm, HttpServletRequest req) {
-        HttpSession session = req.getSession();
-        // The login process should be stopped when there's an active session.
-        if(session.getAttribute("username") != null)
-            return new ResponseEntity("err#activeSession", HttpStatus.OK);
-
-        if(userRepository.existsByUsername(loginForm.getUsername()) == false)
-            return new ResponseEntity("err#wuop", HttpStatus.OK); // wrong username or password
-
-        // compare password
-        User u = userRepository.findByUsername(loginForm.getUsername());
-        boolean isCorrectPassword = BCrypt.checkpw(loginForm.getPassword(), u.getPassword());
-        if(isCorrectPassword == false)
-            return new ResponseEntity("err#wuop", HttpStatus.OK);
-
-        session.setAttribute("username", loginForm.getUsername());
-        return new ResponseEntity("success", HttpStatus.OK);
-    }
-
-    /**
-    *
-    * Handles logout requests.<br>
-    * Active sessions are invalidated and the client is redirected to main page.
-    *
-    */
-    @PostMapping("/logout")
-    @ResponseBody
-    public void logoutUser(HttpServletRequest req, HttpServletResponse res) {
-    	System.out.println("res status: " + res.getStatus());
-        HttpSession session = req.getSession();
-        if(session.getAttribute("username") != null)
-            session.invalidate();
-        
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        if (auth != null) { 
-            new SecurityContextLogoutHandler().logout(req, res, auth);
-        }
-        
-        redirect("/", res);    
-    }
 
     /**
     *
@@ -95,13 +58,13 @@ public class AuthenticationController extends PageControllerEssentials {
         // for first- & lastname fields
         String nameRegex = "^[a-zA-Z]*$";
 
+        // Matches only numbers in a string.
+        String onlyNumbers = "^[0-9]*$";
+
         // Matches (a-zA-Z) (0-9) (. -)
         // allowed username patterns;
-        //      us.er_name
-        //      user_name1
-        //      user.name.1
-        //      1user.name
-        //      user_name_1
+        // us.er_name, user_name1, user.name.1
+        // user.name.1, 1user.name, user_name_1
         String usernameRegex = "^(?![_.])(?!.*[_.]{2})[a-zA-Z0-9._]+(?<![_.])$";
 
         // Trim first & lastname;
@@ -110,6 +73,7 @@ public class AuthenticationController extends PageControllerEssentials {
         registerForm.setLastname(registerForm.getLastname().trim().toLowerCase());
         registerForm.setEmail(registerForm.getEmail().trim().toLowerCase());
         registerForm.setUsername(registerForm.getUsername().trim().toLowerCase());
+        reigsterForm.setPhoneNumber(registerForm.getPhoneNumber().trim());
 
         // Make sure there are no invalid characters ( anything other than a-zA-Z )
         // Only a-zA-Z allowed in first-, lastname fields.
@@ -121,6 +85,9 @@ public class AuthenticationController extends PageControllerEssentials {
 
         if(registerForm.getUsername().matches(usernameRegex) == false)
             return "err#unameInvalid";
+
+        if(registerForm.getPhoneNumber().matches(onlyNumbers) == false)
+            return "err#invalidPhone";
 
         // validate email with apache-commons validator
         // NOTE : returns valid on '€##/&/email@abc.xyz', might need to use regex later.
@@ -139,6 +106,9 @@ public class AuthenticationController extends PageControllerEssentials {
         registerForm.setPassword(pw_hash);
 
         // Everything seems OK !
+
+        // TODO : set default roles & permissions
+
         // register account
         userRepository.save(registerForm);
         System.out.println("Created: " + registerForm.getFirstname());
